# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameArchitecture } from "./gameArchitecture";

export interface AnalysisResult {
  architecture: GameArchitecture;
  warnings: string[];
}