# lib/api-zod/src/generated/types/analyzeRepoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeRepoBodyRole } from "./analyzeRepoBodyRole";
import type { RepoData } from "./repoData";

export interface AnalyzeRepoBody {
  repoData: RepoData;
  /** Whether this repo will contribute graphics or logic to the fusion */
  role: AnalyzeRepoBodyRole;
}