# lib/api-zod/src/generated/types/downloadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FusionResult } from "./fusionResult";

export interface DownloadBody {
  fusionResult: FusionResult;
  gameAName: string;
  gameBName: string;
}