# lib/api-zod/src/generated/types/fileCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileCategoryCategory } from "./fileCategoryCategory";

export interface FileCategory {
  path: string;
  category: FileCategoryCategory;
  reason: string;
  /** @nullable */
  content?: string | null;
}