# lib/api-zod/src/generated/types/fuseGamesBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FuseGamesBodyGameA } from "./fuseGamesBodyGameA";
import type { FuseGamesBodyGameB } from "./fuseGamesBodyGameB";

export interface FuseGamesBody {
  /** Game A providing visuals and world/level design */
  gameA: FuseGamesBodyGameA;
  /** Game B providing game logic and mechanics */
  gameB: FuseGamesBodyGameB;
}