# lib/api-zod/src/generated/types/fusionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedFile } from "./generatedFile";

export interface FusionResult {
  files: GeneratedFile[];
  /** Description of what was combined and how */
  summary: string;
  warnings: string[];
  /** 0-100 score indicating how well the two games could be merged */
  compatibilityScore: number;
}