# lib/api-zod/src/generated/types/gameArchitecture.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileCategory } from "./fileCategory";

export interface GameArchitecture {
  /**
   * Detected rendering engine (canvas2d, three.js, phaser, pixi, webgl, etc.)
   * @nullable
   */
  renderingEngine?: string | null;
  /**
   * Detected game genre
   * @nullable
   */
  gameGenre?: string | null;
  /** High-level description of the game */
  summary: string;
  visualFiles: FileCategory[];
  logicFiles: FileCategory[];
  assetFiles: FileCategory[];
}