# lib/api-zod/src/generated/types/generatedFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedFile {
  path: string;
  content: string;
  description: string;
}