# lib/api-zod/src/generated/types/repoFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RepoFileType } from "./repoFileType";

export interface RepoFile {
  path: string;
  content?: string;
  size: number;
  type: RepoFileType;
}